one server

+++ src/main/java/com/example/learnenglish/repository/LessonRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LessonRepository extends CrudRepository<Lesson, Integer> {
+public interface LessonRepository extends CrudRepository<Lesson, Long> {
 
 }
